+++ biocentral_server/server_management/__init__.py
@@ -2,7 +2,7 @@ from .user_manager import UserManager
 from .process_manager import ProcessManager
 from .file_manager import FileManager, StorageFileType
 from .task_interface import TaskInterface, TaskStatus
-from .embedding_database import init_embeddings_database_instance, EmbeddingsDatabase
+from .embedding_database import init_embeddings_database_instance, EmbeddingsDatabase, EmbeddingsDatabaseTriple
 
 __all__ = [
     'FileManager',
@@ -13,4 +13,5 @@ __all__ = [
     'ProcessManager',
     'init_embeddings_database_instance',
     'EmbeddingsDatabase',
+    'EmbeddingsDatabaseTriple'
 ]
